# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface Reflection {
  id: string;
  comment: string;
  createdAt: string;
}

export interface JournalEntry {
  id: string;
  date: string;
  emotion: string;
  shortAnswer: string;
  longAnswer?: string | null;
  photo?: string | null;
  reflections: Reflection[];
  createdAt: string;
  updatedAt: string;
}

export interface EntryListResponse {
  entries: JournalEntry[];
}

export interface EntryResponse {
  entry: JournalEntry;
}

export interface CreateEntryRequest {
  date: string;
  emotion: string;
  shortAnswer: string;
  longAnswer?: string | null;
  photo?: string | null;
}

export interface UpdateEntryRequest {
  emotion?: string;
  shortAnswer?: string;
  longAnswer?: string | null;
  photo?: string | null;
}

export interface AddReflectionRequest {
  comment: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type BeginBrowserLoginParams = {
  returnTo?: string;
};