# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser authentication endpoints.
  - name: entries
    description: Journal entry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      responses:
        "302":
          description: Redirect after OIDC callback.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session
      responses:
        "302":
          description: Redirect to OIDC logout.

  /entries:
    get:
      tags: [entries]
      operationId: listEntries
      summary: List all entries for the authenticated user
      responses:
        "200":
          description: Entry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"
        "401":
          description: Unauthorized
    post:
      tags: [entries]
      operationId: createEntry
      summary: Create a new journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "401":
          description: Unauthorized

  /entries/{date}:
    get:
      tags: [entries]
      operationId: getEntry
      summary: Get a single entry by date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "404":
          description: Not found
    patch:
      tags: [entries]
      operationId: updateEntry
      summary: Update an entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
    delete:
      tags: [entries]
      operationId: deleteEntry
      summary: Delete an entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /entries/{date}/reflections:
    post:
      tags: [entries]
      operationId: addReflection
      summary: Add a reflection comment to an entry
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReflectionRequest"
      responses:
        "201":
          description: Reflection added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id

    AuthUserEnvelope:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required:
        - user

    Reflection:
      type: object
      properties:
        id:
          type: string
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - comment
        - createdAt

    JournalEntry:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        emotion:
          type: string
        shortAnswer:
          type: string
        longAnswer:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
        reflections:
          type: array
          items:
            $ref: "#/components/schemas/Reflection"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - date
        - emotion
        - shortAnswer
        - reflections
        - createdAt
        - updatedAt

    EntryListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
      required:
        - entries

    EntryResponse:
      type: object
      properties:
        entry:
          $ref: "#/components/schemas/JournalEntry"
      required:
        - entry

    CreateEntryRequest:
      type: object
      properties:
        date:
          type: string
        emotion:
          type: string
        shortAnswer:
          type: string
        longAnswer:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
      required:
        - date
        - emotion
        - shortAnswer

    UpdateEntryRequest:
      type: object
      properties:
        emotion:
          type: string
        shortAnswer:
          type: string
        longAnswer:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true

    AddReflectionRequest:
      type: object
      properties:
        comment:
          type: string
      required:
        - comment

    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error