# lib/api-zod/src/generated/types/createEntryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEntryRequest {
  date: string;
  emotion: string;
  shortAnswer: string;
  longAnswer?: string | null;
  photo?: string | null;
}