# lib/api-zod/src/generated/types/entryListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntry } from "./journalEntry";

export interface EntryListResponse {
  entries: JournalEntry[];
}