# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Reflection } from "./reflection";

export interface JournalEntry {
  id: string;
  date: string;
  emotion: string;
  shortAnswer: string;
  longAnswer?: string | null;
  photo?: string | null;
  reflections: Reflection[];
  createdAt: string;
  updatedAt: string;
}