# lib/api-zod/src/generated/types/reflection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Reflection {
  id: string;
  comment: string;
  createdAt: string;
}